+++ packages/orq-rc/RELEASES.md
@@ -5619,3 +5619,13 @@ Based on:
 - [python v4.3.0-rc.62] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.62] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.62 - packages/orq-rc
+
+## 2026-02-13 10:26:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.715.0 (2.822.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.63] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.63] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.63 - packages/orq-rc